requirements.txt
beautifulsoup4==4.3.2
Django==1.8.2
django-appconf==1.0.1
django-authtools==1.2.0
django-braces==1.8.0
django-compressor==1.5
django-crispy-forms==1.4.0
django-libsass==0.3
django-modelcluster==0.6.2
django-overextends==0.3.2
django-sendfile==0.3.7
django-taggit==0.14.0
django-treebeard==3.0
easy-thumbnails==2.2
html5lib==0.999
libsass==0.8.2
Pillow==2.8.2
psycopg2==2.5.2
pytz==2015.4
requests==2.7.0
six==1.9.0
sphinx-me==0.3
Unidecode==0.4.18
wagtail==1.0rc1
Willow==0.2.1
